Conflicts:
	app/app.iml
	app/build.gradle
	app/src/main/java/com/eveningoutpost/dexdrip/Home.java
	app/src/main/java/com/eveningoutpost/dexdrip/Models/BgReading.java
	app/src/main/java/com/eveningoutpost/dexdrip/Services/DexCollectionService.java
	app/src/main/java/com/eveningoutpost/dexdrip/xdrip.java
	app/src/main/res/xml/pref_data_source.xml

+++ app/src/main/java/com/eveningoutpost/dexdrip/Models/AlertType.java
@@ -12,6 +12,7 @@ import com.activeandroid.annotation.Column;
 import com.activeandroid.annotation.Table;
 import com.activeandroid.query.Select;
 import com.eveningoutpost.dexdrip.Sensor;
+import com.eveningoutpost.dexdrip.UtilityModels.AlertPlayer;
 import com.eveningoutpost.dexdrip.UtilityModels.Notifications;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -33,7 +34,7 @@ public class AlertType extends Model {
     @Column(name = "name")
     public String name;
 
-    @Column(name = "active") 
+    @Column(name = "active")
     public boolean active;
 
     @Column(name = "volume")
@@ -54,7 +55,7 @@ public class AlertType extends Model {
     @Column(name = "time_until_threshold_crossed")
     public double time_until_threshold_crossed;
 
-    // If it is not above, then it must be below. 
+    // If it is not above, then it must be below.
     @Column(name = "above")
     public boolean above;
 
@@ -71,22 +72,25 @@ public class AlertType extends Model {
     public int end_time_minutes;
 
     @Column(name = "minutes_between") //??? what is the difference between minutes_between and default_snooze ???
-    public int minutes_between;
+    public int minutes_between; // The idea here was if ignored it will go off again each x minutes, snooze would be if it was aknowledged and dismissed it will go off again in y minutes
+    // that said, Im okay with doing away with the minutes between and just doing it at a set 5 mins like dex
 
     @Column(name = "default_snooze")
     public int default_snooze;
 
     @Column(name = "text") // ??? what's that? is it different from name?
-    public String text;
-    
+    public String text; // I figured if we wanted some special text, Its
+
     @Column(name = "mp3_file")
     public String mp3_file;
-    
+
     @Column(name = "uuid", index = true)
     public String uuid;
 
+    public final static String LOW_ALERT_55 = "c5f1999c-4ec5-449e-adad-3980b172b920";
     private final static String TAG = Notifications.class.getSimpleName();
-    
+    private final static String TAG_ALERT = "AlertBg";
+
     public static AlertType get_alert(String uuid) {
 
         return new Select()
@@ -94,11 +98,55 @@ public class AlertType extends Model {
         .where("uuid = ? ", uuid)
         .executeSingle();
     }
-    
+
+    /*
+     * This function has 3 needs. In the case of "unclear state" return null
+     * In the case of "unclear state" for more than predefined time, return the "55" alert
+     * In case that alerts are turned off, only return the 55.
+     */
+    public static AlertType get_highest_active_alert(Context context, double bg) {
+
+        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
+        Boolean bg_alerts = prefs.getBoolean("bg_alerts", true);
+        Long UnclearTimeSetting = Long.parseLong(prefs.getString("bg_unclear_readings_minutes", "90")) * 60000;
+
+        Long UnclearTime = BgReading.getUnclearTime(UnclearTimeSetting);
+        AlertType at;
+        if (UnclearTime >= UnclearTimeSetting ) {
+            // we are in an unclear state for too long, ring our alert
+            Log.e(TAG_ALERT, "We are in an unclear state for too long, ring 55 alert");
+            at = get_alert(LOW_ALERT_55);
+            if(at == null) {
+                Log.wtf(TAG, "ERROR, the 55 alert is missing");
+            }
+            return at;
+        }
+        if (UnclearTime > 0) {
+            // We are in an clear state, but not for too long.
+            Log.e(TAG_ALERT, "We are in an clear state, but not for too long. returning null");
+            return null;
+        }
+        at = get_highest_active_alert_helper(bg);
+        if (at != null) {
+            Log.e(TAG_ALERT, "get_highest_active_alert_helper returned alert uuid = " + at.uuid + " alert name = " + at.name);
+        } else {
+            Log.e(TAG_ALERT, "get_highest_active_alert_helper returned NULL");
+        }
+        if (at == null || bg_alerts) {
+            return at;
+        }
+        // alerts are off, return only if this is the 55 alert
+        if(at.uuid.equals(LOW_ALERT_55)) {
+            return at;
+        }
+        return null;
+    }
+
+
     // bg_minute is the estimatin of the bg change rate
-    public static AlertType get_highest_active_alert(double bg, double bg_minute) {
+    private static AlertType get_highest_active_alert_helper(double bg) {
         // Chcek the low alerts
-        
+
         List<AlertType> lowAlerts  = new Select()
             .from(AlertType.class)
             .where("threshold >= ?", bg)
@@ -111,13 +159,13 @@ public class AlertType extends Model {
                 return lowAlert;
             }
         }
-            
+
         // If no low alert found, check higher alert.
         List<AlertType> HighAlerts  = new Select()
             .from(AlertType.class)
             .where("threshold <= ?", bg)
             .where("above = ?", true)
-            .orderBy("threshold asc")
+            .orderBy("threshold desc")
             .execute();
 
         for (AlertType HighAlert : HighAlerts) {
@@ -126,10 +174,10 @@ public class AlertType extends Model {
                 return HighAlert;
             }
         }
-        // no alert found 
+        // no alert found
         return null;
     }
-    
+
     // returns true, if one allert is up and the second is down
     public static boolean OpositeDirection(AlertType a1, AlertType a2) {
         if (a1.above != a2.above) {
@@ -137,7 +185,7 @@ public class AlertType extends Model {
         }
         return false;
     }
-    
+
     // Checks if a1 is more important than a2. returns the higher one
     public static AlertType HigherAlert(AlertType a1, AlertType a2) {
         if (a1.above && !a2.above) {
@@ -164,7 +212,7 @@ public class AlertType extends Model {
             return a2;
         }
     }
-    
+
     public static void remove_all() {
         List<AlertType> Alerts  = new Select()
         .from(AlertType.class)
@@ -173,44 +221,55 @@ public class AlertType extends Model {
         for (AlertType alert : Alerts) {
             alert.delete();
         }
+        ActiveBgAlert.ClearData();
     }
-    
+
     public static void add_alert(
-            String name, 
+            String uuid,
+            String name,
             boolean above,
-            double threshold, 
-            boolean all_day, 
-            int minutes_between, 
+            double threshold,
+            boolean all_day,
+            int minutes_between,
             String mp3_file,
             int start_time_minutes,
             int end_time_minutes,
-            boolean override_silent_mode) {
+            boolean override_silent_mode,
+            int snooze) {
         AlertType at = new AlertType();
         at.name = name;
         at.above = above;
         at.threshold = threshold;
         at.all_day = all_day;
         at.minutes_between = minutes_between;
-        at.uuid = UUID.randomUUID().toString();
+        at.uuid = uuid != null? uuid : UUID.randomUUID().toString();
         at.active = true;
         at.mp3_file = mp3_file;
         at.start_time_minutes = start_time_minutes;
         at.end_time_minutes = end_time_minutes;
         at.override_silent_mode = override_silent_mode;
+        at.default_snooze = snooze;
         at.save();
     }
-    
+
     public static void update_alert(
-            String uuid, 
-            String name, 
-            boolean above, 
-            double threshold, 
-            boolean all_day, 
-            int minutes_between, 
+            String uuid,
+            String name,
+            boolean above,
+            double threshold,
+            boolean all_day,
+            int minutes_between,
             String mp3_file,
             int start_time_minutes,
             int end_time_minutes,
-            boolean override_silent_mode) {
+            boolean override_silent_mode,
+            int snooze) {
+
+        if(uuid.equals(LOW_ALERT_55)) {
+            // This alert can not be removed/updated
+            return;
+        }
+
         AlertType at = get_alert(uuid);
         at.name = name;
         at.above = above;
@@ -223,25 +282,33 @@ public class AlertType extends Model {
         at.start_time_minutes = start_time_minutes;
         at.end_time_minutes = end_time_minutes;
         at.override_silent_mode = override_silent_mode;
+        at.default_snooze = snooze;
         at.save();
     }
     public static void remove_alert(String uuid) {
+        if(uuid.equals(LOW_ALERT_55)) {
+            // This alert can not be removed/updated
+            return;
+        }
         AlertType alert = get_alert(uuid);
-        alert.delete();
+		if(alert != null) {
+	        alert.delete();
+        }
     }
-    
+
     public String toString() {
-        
+
         String name = "name: " + this.name;
         String above = "above: " + this.above;
         String threshold = "threshold: " + this.threshold;
         String all_day = "all_day: " + this.all_day;
-        String minutes_between = "minutes_between: " + this.minutes_between; 
-        String uuid = "uuid: " + this.uuid; 
+        String time = "Start time: " + this.start_time_minutes + " end time: "+ this.end_time_minutes;
+        String minutes_between = "minutes_between: " + this.minutes_between;
+        String uuid = "uuid: " + this.uuid;
 
-        return name + " " + above + " " + threshold + " "+ all_day + " " + minutes_between + " uuid" + uuid;
+        return name + " " + above + " " + threshold + " "+ all_day + " " +time +" " + minutes_between + " uuid" + uuid;
     }
- 
+
     public static void print_all() {
         List<AlertType> Alerts  = new Select()
             .from(AlertType.class)
@@ -252,100 +319,86 @@ public class AlertType extends Model {
             Log.e(TAG, alert.toString());
         }
     }
-    
+
     public static List<AlertType> getAll(boolean above) {
+        String order;
+        if (above) {
+            order = "threshold asc";
+        } else {
+            order = "threshold desc";
+        }
         List<AlertType> alerts  = new Select()
             .from(AlertType.class)
             .where("above = ?", above)
+            .orderBy(order)
             .execute();
 
         return alerts;
     }
-    
-    
-    // This function is a replacment for the UI. It will make sure that there are exactly two alerts
-    // based on what the user has set as high and low. Will be replaced by a UI.
-    public static void CreateStaticAlerts(Context context) {
-        // If there are two alerts already, we are done...
-        List<AlertType> Alerts  = new Select()
-            .from(AlertType.class)
-            .execute();
-        SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-        Double highValue = Double.parseDouble(prefs.getString("highValue", "170"));
-        Double lowValue = Double.parseDouble(prefs.getString("lowValue", "70"));
-        if (Alerts.size() == 2) {
-            if(Alerts.get(0).threshold == highValue && Alerts.get(0).above == true &&
-                    Alerts.get(1).threshold == lowValue && Alerts.get(1).above == false) {
-                Log.e(TAG, "CreateStaticAlerts we have our alerts ok...");
-                return;
-            }
-            if(Alerts.get(1).threshold == highValue && Alerts.get(1).above == true &&
-                    Alerts.get(0).threshold == lowValue && Alerts.get(0).above == false) {
-                Log.e(TAG, "CreateStaticAlerts we have our alerts ok...");
-                return;
-            }
-            
-        }
-        //Log.e(TAG, "CreateStaticAlerts re-creating all our alerts again");
-        //remove_all();
-        //add_alert("high alert", true, highValue, true, 1, null, 0, 0);
-        //add_alert("low alert", false, lowValue, true, 1, null, 0, 0);
-        //print_all();
-    }
-    
-   
-    public static void testAll() {
-        
-        remove_all();
-        add_alert("high alert 1", true, 180, true, 10, null, 0, 0, true);
-        add_alert("high alert 2", true, 200, true, 10, null, 0, 0, true);
-        add_alert("high alert 3", true, 220, true, 10, null, 0, 0, true);
-        print_all();
-        AlertType a1 = get_highest_active_alert(190, 0);
-        Log.e(TAG, "a1 = " + a1.toString());
-        AlertType a2 = get_highest_active_alert(210, 0);
-        Log.e(TAG, "a2 = " + a2.toString());        
 
-        
+
+
+    // This function is used to make sure that we always have a static alert on 55 low.
+    // This alert will not be editable/removable.
+    public static void CreateStaticAlerts() {
+        if(get_alert(LOW_ALERT_55) == null) {
+            add_alert(LOW_ALERT_55, "low alert (unchangable)", false, 55, true, 1, null, 0, 0, true, 20);
+        }
+    }
+
+
+    public static void testAll(Context context) {
+
+        remove_all();
+        add_alert(null, "high alert 1", true, 180, true, 10, null, 0, 0, true, 20);
+        add_alert(null, "high alert 2", true, 200, true, 10, null, 0, 0, true, 20);
+        add_alert(null, "high alert 3", true, 220, true, 10, null, 0, 0, true, 20);
+        print_all();
+        AlertType a1 = get_highest_active_alert(context, 190);
+        Log.e(TAG, "a1 = " + a1.toString());
+        AlertType a2 = get_highest_active_alert(context, 210);
+        Log.e(TAG, "a2 = " + a2.toString());
+
+
         AlertType a3 = get_alert(a1.uuid);
         Log.e(TAG, "a1 == a3 ? need to see true " + (a1==a3) + a1 + " " + a3);
-        
-        add_alert("low alert 1", false, 80, true, 10, null, 0, 0, true);
-        add_alert("low alert 2", false, 60, true, 10, null, 0, 0, true);
-        
-        AlertType al1 = get_highest_active_alert(90, 0);
+
+        add_alert(null, "low alert 1", false, 80, true, 10, null, 0, 0, true, 20);
+        add_alert(null, "low alert 2", false, 60, true, 10, null, 0, 0, true, 20);
+
+        AlertType al1 = get_highest_active_alert(context, 90);
         Log.e(TAG, "al1 should be null  " + al1);
-        al1 = get_highest_active_alert(80, 0);
+        al1 = get_highest_active_alert(context, 80);
         Log.e(TAG, "al1 = " + al1.toString());
-        AlertType al2 = get_highest_active_alert(50, 0);
+        AlertType al2 = get_highest_active_alert(context, 50);
         Log.e(TAG, "al2 = " + al2.toString());
 
         Log.e(TAG, "HigherAlert(a1, a2) = a1?" +  (HigherAlert(a1,a2) == a2));
         Log.e(TAG, "HigherAlert(al1, al2) = al1?" +  (HigherAlert(al1,al2) == al2));
         Log.e(TAG, "HigherAlert(a1, al1) = al1?" +  (HigherAlert(a1,al1) == al1));
         Log.e(TAG, "HigherAlert(al1, a2) = al1?" +  (HigherAlert(al1,a2) == al1));
-        
+
         // Make sure we do not influance on real data...
         remove_all();
-        
+
     }
- 
- 
+
+
     private boolean in_time_frame() {
         if (all_day) {
             //Log.e(TAG, "in_time_frame returning true " );
-            return true; 
+            return true;
         }
         // time_now is the number of minutes that have passed from the start of the day.
         Calendar rightNow = Calendar.getInstance();
         int time_now = toTime(rightNow.get(Calendar.HOUR_OF_DAY), rightNow.get(Calendar.MINUTE));
-        Log.e(TAG, "time_now is " + time_now + " minutes");
+        Log.e(TAG, "time_now is " + time_now + " minutes" + " start_time " + start_time_minutes + " end_time " + end_time_minutes);
         if(start_time_minutes < end_time_minutes) {
             if (time_now >= start_time_minutes && time_now <= end_time_minutes) {
                 return true;
             }
         } else {
-            if (time_now <= start_time_minutes || time_now >= end_time_minutes) {
+            if (time_now >= start_time_minutes || time_now <= end_time_minutes) {
                 return true;
             }
         }
@@ -380,26 +433,54 @@ public class AlertType extends Model {
             return false;
         }
     }
-    
+
+    public static void testAlert(
+        String name,
+        boolean above,
+        double threshold,
+        boolean all_day,
+        int minutes_between,
+        String mp3_file,
+        int start_time_minutes,
+        int end_time_minutes,
+        boolean override_silent_mode,
+        int snooze,
+        Context context) {
+            AlertType at = new AlertType();
+            at.name = name;
+            at.above = above;
+            at.threshold = threshold;
+            at.all_day = all_day;
+            at.minutes_between = minutes_between;
+            at.uuid = UUID.randomUUID().toString();
+            at.active = true;
+            at.mp3_file = mp3_file;
+            at.start_time_minutes = start_time_minutes;
+            at.end_time_minutes = end_time_minutes;
+            at.override_silent_mode = override_silent_mode;
+            at.default_snooze = snooze;
+            new AlertPlayer().getPlayer().startAlert(context, at, "TEST");
+    }
+
     // Time is calculated in minutes. that is 01:20 means 80 minutes.
-    
+
     // This functions are a bit tricky. We can only set time from 00:00 to 23:59 which leaves one minute out. this is because we ignore the
     // seconds. so if the user has set 23:59 we will consider this as 24:00
     // This will be done at the code that reads the time from the ui.
-    
-    
-    
+
+
+
     // return the minutes part of the time
     public static int time2Minutes(int minutes) {
         return (minutes - 60*time2Hours(minutes)) ;
     }
-    
+
  // return the hours part of the time
     public static int time2Hours(int minutes) {
         return minutes / 60;
     }
-    
-    // create the time from hours and minutes. 
+
+    // create the time from hours and minutes.
     public static int toTime(int hours, int minutes) {
         return hours * 60 + minutes;
     }
